fix(ui): extract getBoundingBox helper and memoize bounding box calculation in ProjectMap
